Create ASilentRoom
I have always been into the game - A Dark Room. So I decided to make a game that is based of it, (a knock-off for all you haters out there.) This is still in Alpha but, is open to the public development. If you decide to help contribute to the project, make a GitHub script a put the link in the comments. Thanks!